First pass at a refactor
