+++ functions/src/config.ts
@@ -13,13 +13,50 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
+import { Github } from "./github";
 import * as functions from "firebase-functions";
 
 export class Config {
+  static github = new Github(Config.get("github.token"));
+
+  static readonly FEATURED_BLACKLIST_PROJECTS_URL = Github.getRawContentUrl(
+    "firebase",
+    "firebaseopensource.com",
+    "config/feature_blacklist_projects.json"
+  );
+
+  static readonly ADDITIONAL_PROJECTS_URL = Github.getRawContentUrl(
+    "firebase",
+    "firebaseopensource.com",
+    "config/additional_projects.json"
+  );
+
+  static readonly ADDITIONAL_PROJECTS: string[] = [];
+  static readonly FEATURED_BLACKLIST_PROJECTS: string[] = [];
+
+  static loadGlobalConfig = async function() {
+    if (
+      this.ADDITIONAL_PROJECTS.length > 0 &&
+      this.FEATURED_BLACKLIST_PROJECTS.length > 0
+    ) {
+      return;
+    }
+
+    const additionalData = await this.github.getContent(
+      this.ADDITIONAL_PROJECTS_URL
+    );
+    this.ADDITIONAL_PROJECTS = JSON.parse(additionalData).projects;
+
+    const blacklistData = await this.github.getContent(
+      this.FEATURED_BLACKLIST_PROJECTS_URL
+    );
+    this.FEATURED_BLACKLIST_PROJECTS = JSON.parse(blacklistData).projects;
+  };
+
   /**
    * Get a config key, either from the env or from local.
    */
-  get(key: string): string {
+  static get(key: string): string {
     try {
       return this._getNestedProperty(functions.config(), key);
     } catch (e) {
@@ -28,7 +65,7 @@ export class Config {
     }
   }
 
-  _getNestedProperty(obj: any, name: string): string {
+  static _getNestedProperty(obj: any, name: string): string {
     const pieces = name.split(".");
     let val = obj;
     pieces.forEach(piece => {
