+++ functions/src/test/test.ts
@@ -18,6 +18,7 @@
 //
 // To run:
 // GCLOUD_PROJECT="fir-oss" github_token="<YOUR_GITHUB_TOKEN>" node test/test.js
+import { Config } from "../config";
 import { Project } from "../project";
 
 const project = new Project();
@@ -35,7 +36,7 @@ const ids = [
   "googlesamples::easypermissions"
 ];
 
-project.loadGlobalConfig().then(() => {
+Config.loadGlobalConfig().then(() => {
   ids.forEach(id => {
     return project.recursiveStoreProject(id).catch(err => {
       console.warn(err);
